feat: replace login with SIWE
