+++ client/App.tsx
@@ -6,7 +6,6 @@ import axios from '@/lib/axios';
 import Loading from './pages/Loading';
 import Login from './pages/Login';
 import Register from './pages/Register';
-import LoginSIWE from './pages/LoginSIWE';
 import Navbar from './components/Navbar';
 import Home from './pages/buyer/Home';
 import Catalog from './pages/buyer/Catalog';
@@ -68,7 +67,6 @@ function App() {
             <Routes>
               <Route path="/" element={<Loading />} />
               <Route path="/login" element={<Login />} />
-              <Route path="/login-siwe" element={<LoginSIWE />} />
               <Route path="/register" element={<Register />} />
               <Route path="/dashboard" element={<Dashboard />} />
