+++ client/pages/Login.tsx
@@ -1,70 +1,125 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Input } from '../components/ui/input';
-import { Label } from '../components/ui/label';
-import { Button } from '../components/ui/button';
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-} from '../components/ui/card';
-import { Alert, AlertDescription } from '../components/ui/alert';
-import { Badge } from '../components/ui/badge';
-import axios from '@/lib/axios';
 import { useAtom, useSetAtom } from 'jotai';
-import { tokenAtom, userAtom } from '../atoms/loginAtoms';
+import { tokenAtom, userAtom } from '@/atoms/loginAtoms';
 import { loadWalletAtom } from '@/atoms/walletAtoms';
-import { Store, AlertCircle, CheckCircle2, Loader2 } from 'lucide-react';
+import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { Input } from '@/components/ui/input';
+import { Alert, AlertDescription } from '@/components/ui/alert';
+import { AlertCircle, Loader2 } from 'lucide-react';
+import axios from '@/lib/axios';
+import {
+  createAppKit,
+  useAppKit,
+  useAppKitAccount,
+  useAppKitNetworkCore,
+  useAppKitProvider,
+} from '@reown/appkit/react';
+import { SiweMessage } from 'siwe';
+import { EthersAdapter } from '@reown/appkit-adapter-ethers';
+import { arbitrum, mainnet } from '@reown/appkit/networks';
+import { BrowserProvider, JsonRpcSigner, type Eip1193Provider } from 'ethers';
+
+// Initialize AppKit once
+// createAppKit({
+//   projectId: '3dc8fb97b90a536ce400e5d65a3f5ff8',
+//   networks: [mainnet],
+//   defaultNetwork: mainnet,
+// features: {
+//   email: true, // default to true
+//   socials: [
+//     "google",
+//     "x",
+//     "github",
+//     "discord",
+//     "apple",
+//     "facebook",
+//     "farcaster",
+//   ],
+//   emailShowWallets: true, // default to true
+// },
+// })
+
+createAppKit({
+  adapters: [new EthersAdapter()],
+  networks: [mainnet, arbitrum],
+  projectId: '3dc8fb97b90a536ce400e5d65a3f5ff8',
+  defaultAccountTypes: { eip155: 'eoa' },
+  enableNetworkSwitch: false,
+  features: {
+    email: false,
+    socials: ['google', 'facebook'],
+    emailShowWallets: false, // default to true
+  },
+});
 
 function Login() {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
   const navigate = useNavigate();
-  const [token] = useAtom(tokenAtom);
   const [, setToken] = useAtom(tokenAtom);
   const [, setUser] = useAtom(userAtom);
   const loadWallet = useSetAtom(loadWalletAtom);
+  const { open } = useAppKit();
+  const { address, isConnected } = useAppKitAccount();
+  const { walletProvider } = useAppKitProvider<unknown>('eip155');
+  const [error, setError] = useState<string | null>(null);
+  const [isLoading, setIsLoading] = useState(false);
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError(null);
+  // const { address, isConnected } = useAppKitAccount();
+  const { chainId } = useAppKitNetworkCore();
+  // const { walletProvider } = useAppKitProvider('eip155');
+
+  const handleConnect = () => {
+    open();
+  };
+
+  const handleSignIn = async () => {
+    if (!walletProvider || !address) return;
     setIsLoading(true);
+    setError(null);
 
     try {
-      let res = await axios.get('/api/settings');
-      console.log(res.data);
-      res = await axios.post('/api/login', { username, password });
-      console.log(res.data);
-      const token = res.data.token;
+      const nonce = Math.random().toString(36).substring(2, 10);
+      const cfg = {
+        domain: window.location.host,
+        address,
+        statement: 'Sign in with Ethereum to MarketPlace',
+        uri: window.location.origin,
+        version: '1',
+        chainId: chainId,
+        nonce,
+      };
 
-      // Set token in atoms first so axios interceptor can use it
-      setToken(token);
+      const msg = new SiweMessage(cfg as Partial<SiweMessage>);
+      const message = msg.prepareMessage();
+      const signerProvider = new BrowserProvider(
+        walletProvider as unknown as Eip1193Provider,
+        Number(chainId)
+      );
+      const signer = new JsonRpcSigner(signerProvider, address);
+      const signature = await signer.signMessage(message);
 
-      res = await axios.get('/api/me');
-      console.log(res.data);
-      const user = res.data;
+      // const signature = await (
+      //   walletProvider as unknown as {
+      //     request: (args: {
+      //       method: string
+      //       params: unknown[]
+      //     }) => Promise<string>
+      //   }
+      // ).request({
+      //   method: 'personal_sign',
+      //   params: [message, address],
+      // })
 
-      // Set user in atoms
-      setUser(user);
+      const res = await axios.post('/api/login/siwe', { message, signature });
+      setToken(res.data.token);
+      setUser(res.data.user);
       await loadWallet();
-
-      if (user.role === 'admin') {
-        console.log('navigate..');
-        navigate('/admin', { replace: true });
-      } else if (user.role === 'buyer') {
-        navigate('/home', { replace: true });
-      } else if (user.role === 'seller') {
-        navigate('/seller', { replace: true });
-      }
+      navigate('/home', { replace: true });
     } catch (err) {
-      if (err instanceof Error) {
-        setError(err.message);
-      } else {
-        setError('Failed to login');
-      }
+      console.log('!!!');
+      const msg = err instanceof Error ? err.message : 'Login failed';
+      setError(msg);
     } finally {
       setIsLoading(false);
     }
@@ -72,109 +127,56 @@ function Login() {
 
   return (
     <div className="flex items-center justify-center p-4 bg-gradient-to-br from-background to-muted/20">
-      <div className="w-full max-w-md space-y-8">
-        {/* Brand Header */}
-        <div className="text-center space-y-3">
-          <div className="flex items-center justify-center space-x-2">
-            <div className="p-2 bg-primary/10 rounded-lg">
-              <Store className="h-6 w-6 text-primary" />
-            </div>
-            <span className="font-bold text-2xl">MarketPlace</span>
-          </div>
-          <p className="text-sm text-muted-foreground">
-            The digital trade nexus
-          </p>
-          <h2 className="text-xl font-semibold text-foreground">
-            Sign in to your account
-          </h2>
-        </div>
+      <Card className="w-full max-w-md shadow-lg border-0">
+        <CardHeader>
+          <CardTitle className="text-xl text-center">
+            Sign In with Ethereum
+          </CardTitle>
+        </CardHeader>
+        <CardContent className="space-y-4">
+          <Button
+            className="w-full"
+            onClick={handleConnect}
+            disabled={isLoading}
+          >
+            Connect Wallet
+          </Button>
 
-        {/* Login Form */}
-        <Card className="shadow-lg border-0">
-          <CardHeader className="space-y-3 pb-6">
-            <CardTitle className="text-xl text-center">Welcome back</CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            <form onSubmit={handleSubmit} className="space-y-4">
-              <div className="space-y-2">
-                <Label htmlFor="username" className="text-sm font-medium">
-                  Username
-                </Label>
-                <Input
-                  id="username"
-                  value={username}
-                  onChange={(e) => setUsername(e.target.value)}
-                  autoComplete="username"
-                  placeholder="Enter your username"
-                  disabled={isLoading}
-                  className="h-10"
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="password" className="text-sm font-medium">
-                  Password
-                </Label>
-                <Input
-                  id="password"
-                  type="password"
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  autoComplete="current-password"
-                  placeholder="Enter your password"
-                  disabled={isLoading}
-                  className="h-10"
-                />
-              </div>
-
-              {error && (
-                <Alert variant="destructive">
-                  <AlertCircle className="h-4 w-4" />
-                  <AlertDescription>{error}</AlertDescription>
-                </Alert>
-              )}
-
-              {token && (
-                <Alert>
-                  <CheckCircle2 className="h-4 w-4" />
-                  <AlertDescription>
-                    <div className="flex items-center gap-2">
-                      <span>Logged in successfully</span>
-                      <Badge variant="secondary" className="text-xs">
-                        Token saved
-                      </Badge>
-                    </div>
-                  </AlertDescription>
-                </Alert>
-              )}
-
+          {error && (
+            <Alert variant="destructive">
+              <AlertCircle className="h-4 w-4" />
+              <AlertDescription>{error}</AlertDescription>
+            </Alert>
+          )}
+          {isConnected ? (
+            <div className="space-y-4">
+              <Input readOnly value={address} />
               <Button
-                type="submit"
-                className="w-full h-10"
+                className="w-full"
+                onClick={handleSignIn}
                 disabled={isLoading}
               >
                 {isLoading ? (
                   <>
-                    <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                    Signing in...
+                    <Loader2 className="mr-2 h-4 w-4 animate-spin" /> Signing
+                    in...
                   </>
                 ) : (
                   'Sign In'
                 )}
               </Button>
-            </form>
-          </CardContent>
-        </Card>
-        <p className="text-center text-sm mt-4">
-          Don't have an account?{' '}
-          <button
-            type="button"
-            className="underline"
-            onClick={() => navigate('/register')}
-          >
-            Register
-          </button>
-        </p>
-      </div>
+            </div>
+          ) : (
+            <Button
+              className="w-full"
+              onClick={handleConnect}
+              disabled={isLoading}
+            >
+              Connect Wallet
+            </Button>
+          )}
+        </CardContent>
+      </Card>
     </div>
   );
 }
