+++ package.json
@@ -65,10 +65,15 @@
   "devDependencies": {
     "@eslint/js": "^9.30.1",
     "@stoplight/spectral": "^6.5.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/jest": "^29.5.11",
     "@types/node": "^24.0.10",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
     "@types/react-router-dom": "^5.3.3",
+    "@types/react-test-renderer": "^19.1.0",
     "@types/sql.js": "^1.4.9",
     "@vitejs/plugin-react": "^4.5.2",
     "eslint": "^9.29.0",
@@ -76,14 +81,15 @@
     "eslint-plugin-react-refresh": "^0.4.20",
     "gh-pages": "^6.3.0",
     "globals": "^16.2.0",
+    "jest-environment-jsdom": "^30.0.4",
     "msw": "^1.3.5",
     "prettier": "^3.6.2",
     "react-router-dom": "^7.6.3",
+    "react-test-renderer": "^19.1.0",
     "speccy": "^0.11.0",
+    "ts-jest": "^29.4.0",
     "tw-animate-css": "^1.3.5",
     "typescript": "~5.8.3",
-    "ts-jest": "^29.4.0",
-    "@types/jest": "^29.5.11",
     "typescript-eslint": "^8.34.1",
     "vite": "^7.0.0"
   },
