+++ tests/appRouting.test.tsx
@@ -1,0 +1,4 @@
+
+test.skip('Routing placeholder', () => {
+  expect(true).toBe(true);
+});
