+++ tests/loginPage.test.tsx
@@ -1,0 +1,4 @@
+
+test.skip('Login component placeholder', () => {
+  expect(true).toBe(true);
+});
